# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareOS API — home care operating system foundation.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: caregiverId
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleCreateResult"
  /schedules/{id}:
    patch:
      operationId: updateSchedule
      tags: [schedules]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /pay-periods/{id}:
    get:
      operationId: getPayPeriod
      tags: [payroll]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayPeriodDetail"
  /pay-periods/{id}/close:
    post:
      operationId: closePayPeriod
      tags: [payroll]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayPeriod"
  /pay-periods/{id}/export:
    get:
      operationId: exportPayPeriodCsv
      tags: [payroll]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ADP-compatible CSV
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    ScheduleStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, COMPLETED, MISSED, CANCELLED]
    Schedule:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        caregiverId: { type: string }
        caregiverName: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        scheduledMinutes: { type: integer }
        serviceCode: { type: string }
        serviceDescription: { type: string }
        authorizationId: { type: string, nullable: true }
        status:
          $ref: "#/components/schemas/ScheduleStatus"
        notes: { type: string, nullable: true }
      required:
        [
          id,
          clientId,
          clientName,
          caregiverId,
          caregiverName,
          startTime,
          endTime,
          scheduledMinutes,
          serviceCode,
          serviceDescription,
          status,
        ]
    CreateScheduleBody:
      type: object
      properties:
        clientId: { type: string }
        caregiverId: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        serviceCode: { type: string }
        serviceDescription: { type: string }
        authorizationId: { type: string }
        notes: { type: string }
      required: [clientId, caregiverId, startTime, endTime, serviceCode]
    UpdateScheduleBody:
      type: object
      properties:
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        caregiverId: { type: string }
        status:
          $ref: "#/components/schemas/ScheduleStatus"
        notes: { type: string }
    ScheduleConflict:
      type: object
      properties:
        type:
          type: string
          enum:
            [
              DOUBLE_BOOK,
              EXPIRED_CREDENTIAL,
              AUTH_OVERRUN,
              AUTH_EXPIRED,
              OT_THRESHOLD,
            ]
        message: { type: string }
        severity:
          type: string
          enum: [WARNING, BLOCK]
      required: [type, message, severity]
    ScheduleCreateResult:
      type: object
      properties:
        schedule:
          $ref: "#/components/schemas/Schedule"
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleConflict"
      required: [schedule, conflicts]
...
    PayPeriodStatus:
      type: string
      enum: [OPEN, CLOSED, EXPORTED, PAID]
    PayPeriod:
      type: object
      properties:
        id: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        status:
          $ref: "#/components/schemas/PayPeriodStatus"
        totalRegularHours: { type: number }
        totalOvertimeHours: { type: number }
        totalDoubleTimeHours: { type: number }
        totalGrossPay: { type: number }
        caregiverCount: { type: integer }
        exportedAt: { type: string, format: date-time, nullable: true }
      required:
        [
          id,
          startDate,
          endDate,
          status,
          totalRegularHours,
          totalOvertimeHours,
          totalDoubleTimeHours,
          totalGrossPay,
          caregiverCount,
        ]
    TimeEntry:
      type: object
      properties:
        id: { type: string }
        caregiverId: { type: string }
        caregiverName: { type: string }
        visitId: { type: string }
        workDate: { type: string, format: date }
        regularMinutes: { type: integer }
        overtimeMinutes: { type: integer }
        doubleTimeMinutes: { type: integer }
        payRate: { type: number }
        regularPay: { type: number }
        overtimePay: { type: number }
        doubleTimePay: { type: number }
        totalPay: { type: number }
        ruleEngineVersion: { type: string }
      required:
        [
          id,
          caregiverId,
          caregiverName,
          visitId,
          workDate,
          regularMinutes,
          overtimeMinutes,
          doubleTimeMinutes,
          payRate,
          regularPay,
          overtimePay,
          doubleTimePay,
          totalPay,
          ruleEngineVersion,
        ]
    PayPeriodDetail:
      allOf:
        - $ref: "#/components/schemas/PayPeriod"
        - type: object
          properties:
            entries:
              type: array
              items:
                $ref: "#/components/schemas/TimeEntry"
            byCaregiver:
              type: array
              items:
                $ref: "#/components/schemas/CaregiverPayrollSummary"
          required: [entries, byCaregiver]
    CaregiverPayrollSummary:
      type: object
      properties:
        caregiverId: { type: string }
        caregiverName: { type: string }
        regularMinutes: { type: integer }
        overtimeMinutes: { type: integer }
        doubleTimeMinutes: { type: integer }
        regularPay: { type: number }
        overtimePay: { type: number }
        doubleTimePay: { type: number }
        totalPay: { type: number }
      required:
        [
          caregiverId,
          caregiverName,
          regularMinutes,
          overtimeMinutes,
          doubleTimeMinutes,
          regularPay,
          overtimePay,
          doubleTimePay,
          totalPay,
        ]
...